restructure eslint-config
