+++ packages/eslint-config/core/bestpractices.js
