+++ packages/eslint-config/core/errors.js
