+++ packages/eslint-config/core/es6.js
