+++ packages/eslint-config/core/plugin-jsdoc.js
