+++ packages/eslint-config/core/style.js
