+++ packages/eslint-config/core/variables.js
