+++ packages/eslint-config/package.json
@@ -34,7 +34,7 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/FuelRats/eslint-config-fuelrats.git",
-    "directory": "eslint-config"
+    "directory": "packages/eslint-config"
   },
   "homepage": "https://github.com/FuelRats/eslint-config-fuelrats",
   "keywords": [
@@ -48,7 +48,7 @@
   "main": "index.js",
   "scripts": {
     "release": "yarn publish --non-interactive",
-    "release:canary": "yarn publish --tag rc --non-interactive"
+    "release:canary": "yarn publish --tag canary --non-interactive"
   },
   "dependencies": {
     "eslint-restricted-globals": "^0.2.0"
