+++ packages/eslint-config/purejs.js
@@ -8,13 +8,13 @@ module.exports = {
     es6: true,
   },
   extends: [
-    './rules/bestpractices',
-    './rules/errors',
-    './rules/es6',
-    './rules/style',
-    './rules/variables',
-    './rules/plugin-import',
-    './rules/plugin-jsdoc',
+    './core/bestpractices',
+    './core/errors',
+    './core/es6',
+    './core/style',
+    './core/variables',
+    './core/plugin-import',
+    './core/plugin-jsdoc',
   ].map(require.resolve),
   parserOptions: {
     ecmaVersion: 2020,
